locaiver-back/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
autopep8==2.3.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
docstring_parser==0.17.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.111.0
google-cloud-bigquery==3.36.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.32.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
idna==3.10
numpy==2.3.2
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
flask==3.1.2
gunicorn==21.2.0
imageio-ffmpeg==0.5.1
imageio[ffmpeg]
moviepy==1.0.3